+++ components/Core/include/Core/TypeModelConverter.hpp
@@ -46,7 +46,7 @@ private:
     std::shared_ptr<ifdk_objects::type::Type> createTask();
     std::shared_ptr<ifdk_objects::type::Type> createCore();
     std::shared_ptr<ifdk_objects::type::Type> createGateway(const std::string &name);
-    std::shared_ptr<ifdk_objects::type::Type> createModule(uint32_t id);
+    std::shared_ptr<ifdk_objects::type::Type> createModule(uint16_t id);
     std::shared_ptr<ifdk_objects::type::Type> createLogService();
 
     /** Get system characteristics */
