+++ components/Core/src/TypeModelConverter.cpp
@@ -51,7 +51,7 @@ std::shared_ptr<TypeModel> TypeModelConverter::createModel()
     addSubsystemSubType(typeMap, createCore());
 
     /* modules */
-    for (uint32_t moduleId = 0; moduleId < mSystem.getModuleEntries().size(); ++moduleId) {
+    for (uint16_t moduleId = 0; moduleId < mSystem.getModuleEntries().size(); ++moduleId) {
         addSubsystemSubType(typeMap, createModule(moduleId));
     }
 
@@ -131,7 +131,7 @@ std::shared_ptr<Type> TypeModelConverter::createSubsystem()
     /* Modules*/
     auto compColl = std::make_shared<ComponentRefCollection>(collectionName_module);
 
-    for (uint32_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
+    for (uint16_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
         std::string moduleName = findModuleEntryName(i);
 
         compColl->add(ComponentRef(moduleName));
@@ -153,7 +153,7 @@ std::shared_ptr<Type> TypeModelConverter::createPipe()
     /* Modules*/
     auto compColl = std::make_shared<ComponentRefCollection>(collectionName_module);
 
-    for (uint32_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
+    for (uint16_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
         std::string moduleName = findModuleEntryName(i);
 
         compColl->add(ComponentRef(moduleName));
@@ -178,7 +178,7 @@ std::shared_ptr<Type> TypeModelConverter::createTask()
     /* Modules*/
     auto compColl = std::make_shared<ComponentRefCollection>(collectionName_module);
 
-    for (uint32_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
+    for (uint16_t i = 0; i < mSystem.getModuleEntries().size(); ++i) {
         std::string moduleName = findModuleEntryName(i);
 
         compColl->add(ComponentRef(moduleName));
@@ -213,7 +213,7 @@ std::shared_ptr<Type> TypeModelConverter::createGateway(const std::string &name)
     return gateway;
 }
 
-std::shared_ptr<Type> TypeModelConverter::createModule(uint32_t id)
+std::shared_ptr<Type> TypeModelConverter::createModule(uint16_t id)
 {
     /* Asserting because id is the result of iteration in [0..mSystem.getModuleEntries().size()[*/
     assert(id < mSystem.getModuleEntries().size());
